walletProject/requirements.txt
Django==1.11.14
django-rest-auth==0.9.3
djangorestframework==3.8.2
pkg-resources==0.0.0
pytz==2018.5
six==1.11.0
